backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
python-multipart
passlib[bcrypt]
python-jose[cryptography]
geoip2 